Thêm component thông báo, Topbar để hiển thị tài khoản, thêm cam kết và hiển thị nội dung cam kết ở phiếu thu

+++ src/services/DanhMuc/chucVuService.ts
@@ -49,6 +49,14 @@ export const capNhatChucVu = async (ma_chuc_vu: number, data: ChucVuData) => {
 }
 
 export const xoaChucVu = async (ma_chuc_vu: number) => {
+    const countNhanSu = await prisma.nhanSu.count({
+        where: { ma_chuc_vu: ma_chuc_vu },
+    })
+
+    if (countNhanSu > 0) {
+        throw new Error(`CONFLICT_RELATION`)
+    }
+
     return await prisma.chucVu.delete({
         where: { ma_chuc_vu: ma_chuc_vu },
     })
